Changed NWM input file source.
The NWM stream network file gets downloaded and cached from the URL
provided in the National Water Model website (https://water.noaa.gov/about/nwm).
The structure of this new gdb file is different from the original, but
contains the same information that the program requires as the original.
The outputs need to be validated.